pyproject.toml
[tool.poetry]
name = "python_project_template"
version = "0.1.0"
description = ""
authors = ["davidtnfsh <davidtnfsh@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7.1"
requests = "^2.23.0"
apache-airflow = {extras = ["google_auth"], version = "^1.10.13"}
searchconsole = "^0.0.4"
google-cloud-bigquery = "<2.8.0"
pycontw-report-generator = "^1.5.0"
pandas = "^1.2.3"
pyarrow = "^3.0.0"

[tool.poetry.dev-dependencies]
safety = "^1.9.0"
bandit = "^1.6.2"
pytest-cov = "^2.8.1"
black = "^19.10b0"
mypy = "^0.770"
pytest = "^5.4.1"
flake8 = "^3.8.3"
isort = "^5.6.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"


[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''


[tool.isort]
profile = "black"


[tool.coverage]
    [tool.coverage.report]
    show_missing = true
    exclude_lines = [
        # Have to re-enable the standard pragma
        'pragma: no cover',

        # Don't complain about missing debug-only code:
        'def __repr__',
        'if self\.debug',

        # Don't complain if tests don't hit defensive assertion code:
        'raise AssertionError',
        'raise NotImplementedError',

        # Don't complain if non-runnable code isn't run:
        'if 0:',
        'if __name__ == .__main__.:'
    ]
    omit = [
        'env/*',
        'venv/*',
        '*/virtualenv/*',
        '*/virtualenvs/*',
        '*/tests/*'
    ]
